refactor: move deep link methods to own lib
