+++ packages/shared/lib/notifications.ts
@@ -1,9 +1,8 @@
-import { appSettings } from 'shared/lib/appSettings'
-import { generateRandomId } from 'shared/lib/utils'
-import { get, writable } from 'svelte/store'
-import { Electron } from './electron'
-
-import type { NotificationData, NotificationType } from './typings/notification'
+import { appSettings } from 'shared/lib/appSettings';
+import { generateRandomId } from 'shared/lib/utils';
+import { get, writable } from 'svelte/store';
+import { Electron } from './electron';
+import type { NotificationData, NotificationType } from './typings/notification';
 
 const NOTIFICATION_TIMEOUT_DEFAULT = 5000
 export const NOTIFICATION_TIMEOUT_NEVER = -1
